Se agrego clase response para formatear el json de respuesta, tambien se modifico los metodos del controlador para tener un funcionamiento correcto y la modificación del services de Restaurant tambien se realizo

+++ src/main/java/com/luisedev/restaurant/services/RestaurantService.java
@@ -1,6 +1,5 @@
 package com.luisedev.restaurant.services;
 
-import com.google.zxing.qrcode.encoder.QRCode;
 import com.luisedev.restaurant.dto.RestaurantCaractDto;
 import com.luisedev.restaurant.dto.RestaurantDto;
 import com.luisedev.restaurant.mapper.MapperRestaurant;
@@ -8,18 +7,16 @@ import com.luisedev.restaurant.model.Restaurant;
 import com.luisedev.restaurant.model.RestaurantCaract;
 import com.luisedev.restaurant.repository.RestaurantCaractRepository;
 import com.luisedev.restaurant.repository.RestaurantRepository;
-import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Base64;
+
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Service
 public class RestaurantService implements IRestaurantService{
-
-    @Autowired
-    private QRCode qrCodeService;
 
     @Autowired
     private RestaurantRepository restaurantRepository;
@@ -28,47 +25,97 @@ public class RestaurantService implements IRestaurantService{
 
 
     @Override
-    public List<RestaurantDto> getAllRestaurant() {
-        return restaurantRepository.findAll().stream().map(r -> {
-            return RestaurantDto.builder()
-                    .id(r.getId())
-                    .ruc(r.getRuc())
-                    .street(r.getStreet())
-                    .district(r.getDistrict())
-                    .logoPhoto(r.getLogoPhoto())
-                    .caractDto(RestaurantCaractDto.builder()
-                            .id(r.getIdCaractRestaurant().getId())
-                            .qrCode(r.getIdCaractRestaurant().getQr_code())
-                            .tableQty(r.getIdCaractRestaurant().getTableQty())
-                            .chairTable(r.getIdCaractRestaurant().getChairTable())
-                            .floorQty(r.getIdCaractRestaurant().getFloorQty())
-                            .build())
-                    .build();
-        }).toList();
-    }
-
-
-    @Override
-    public RestaurantDto createRestaurant(RestaurantDto restaurantDto) {
-
-        return MapperRestaurant.restaurantDto(restaurantRepository.save(Restaurant.builder()
-                        .ruc(restaurantDto.getRuc())
-                        .district(restaurantDto.getDistrict())
-                        .street(restaurantDto.getLogoPhoto())
-                        .razonsocial(restaurantDto.getRazonSocial())
-                        .logoPhoto(restaurantDto.getLogoPhoto())
-                        .idCaractRestaurant(restaurantCaractRepository.save(RestaurantCaract.builder()
-                                        .chairTable(restaurantDto.getCaractDto().getChairTable())
-                                        .qr_code(restaurantDto.getCaractDto().getQrCode())
-                                        .tableQty(restaurantDto.getCaractDto().getTableQty())
-                                        .floorQty(restaurantDto.getCaractDto().getFloorQty())
-                                .build()))
-                .build()));
-    }
-
-    @Override
-    public RestaurantDto editRestaurant(Long id, RestaurantDto restaurantDto) {
+    public Response  getAllRestaurant() {
         try {
+            HashMap<String, Object> json = new HashMap<>();
+
+            List<RestaurantDto> restaurantDtoList = restaurantRepository.findAll().stream().map(r -> {
+                return RestaurantDto.builder()
+                        .id(r.getId())
+                        .ruc(r.getRuc())
+                        .street(r.getStreet())
+                        .district(r.getDistrict())
+                        .logoPhoto(r.getLogoPhoto())
+                        .caractDto(RestaurantCaractDto.builder()
+                                .id(r.getIdCaractRestaurant().getId())
+                                .qrCode(r.getIdCaractRestaurant().getQr_code())
+                                .tableQty(r.getIdCaractRestaurant().getTableQty())
+                                .chairTable(r.getIdCaractRestaurant().getChairTable())
+                                .floorQty(r.getIdCaractRestaurant().getFloorQty())
+                                .build())
+                        .build();
+            }).toList();
+
+            json.put("message", "All the restaurant actuality");
+            json.put("body", restaurantDtoList);
+
+
+
+            return Response.builder()
+                    .status(200)
+                    .success(true)
+                    .data(json)
+                    .build();
+        } catch (Exception e) {
+            HashMap<String, Object> json = new HashMap<>();
+
+            json.put("message", "Error in the visualization the restaurant");
+            json.put("body", e.getMessage());
+
+            return Response.builder()
+                    .status(500)
+                    .success(true)
+                    .data(json)
+                    .build();
+        }
+
+    }
+
+
+    @Override
+    public Response createRestaurant(RestaurantDto restaurantDto) {
+        try {
+            HashMap<String, Object> json = new HashMap<>();
+            RestaurantDto restaurant = MapperRestaurant.restaurantDto(restaurantRepository.save(Restaurant.builder()
+                    .ruc(restaurantDto.getRuc())
+                    .district(restaurantDto.getDistrict())
+                    .street(restaurantDto.getLogoPhoto())
+                    .razonsocial(restaurantDto.getRazonSocial())
+                    .logoPhoto(restaurantDto.getLogoPhoto())
+                    .idCaractRestaurant(restaurantCaractRepository.save(RestaurantCaract.builder()
+                            .chairTable(restaurantDto.getCaractDto().getChairTable())
+                            .qr_code(restaurantDto.getCaractDto().getQrCode())
+                            .tableQty(restaurantDto.getCaractDto().getTableQty())
+                            .floorQty(restaurantDto.getCaractDto().getFloorQty())
+                            .build()))
+                    .build()));
+            json.put("message", "Restaurant created correctly");
+            json.put("body", restaurant);
+
+            return Response.builder()
+                    .status(201)
+                    .success(true)
+                    .data(json)
+                    .build();
+        } catch (Exception e) {
+
+            HashMap<String, Object> json = new HashMap<>();
+            json.put("message", "Error in the moment created Restaurant");
+            json.put("body", e.getMessage());
+
+            return Response.builder()
+                    .success(true)
+                    .status(500)
+                    .data(json)
+                    .build();
+        }
+
+    }
+
+    @Override
+    public Response editRestaurant(Long id, RestaurantDto restaurantDto) {
+        try {
+            HashMap<String, Object> json = new HashMap<>();
             Restaurant restaurant = restaurantRepository.findById(id).orElseThrow(() -> new RuntimeException("Restaurant with this id not exits"));
             restaurant.setDistrict(restaurantDto.getDistrict());
             restaurant.setRuc(restaurantDto.getRazonSocial());
@@ -79,26 +126,58 @@ public class RestaurantService implements IRestaurantService{
             restaurant.getIdCaractRestaurant().setQr_code(restaurantDto.getCaractDto().getQrCode());
             restaurant.getIdCaractRestaurant().setTableQty(restaurantDto.getCaractDto().getTableQty());
 
-            return MapperRestaurant.restaurantDto(restaurantRepository.save(restaurant));
+            restaurant = restaurantRepository.save(restaurant);
+
+            json.put("message", "Restaurant has ben edited correctly");
+            json.put("body", restaurant);
+
+
+            return Response.builder()
+                    .status(200)
+                    .success(true)
+                    .data(json)
+                    .build();
 
         } catch (Exception e) {
-            throw new RuntimeException(e.getMessage());
+            HashMap<String, Object> json = new HashMap<>();
+            json.put("message", "Error in the moment the edited Restaurant");
+            json.put("body", e.getMessage());
+
+            return Response.builder()
+                    .success(true)
+                    .status(500)
+                    .data(json)
+                    .build();
         }
 
     }
 
     @Override
-    public String deleteRestaurant(Long id, RestaurantDto restaurantDto) {
+    public Response deleteRestaurant(Long id) {
         try {
+            Map<String, Object> json = new HashMap<>();
             Restaurant restaurant = restaurantRepository.findById(id).orElseThrow(() -> new RuntimeException("Restaurant with this id not exits"));
-            restaurantRepository.deleteById(id);
-            return "That restaurant has been removed";
+            if (restaurant != null) {
+                restaurantRepository.deleteById(id);
+                json.put("message", "Restaurant has been deleted correctly");
+                json.put("body", null);
+                return  Response.builder()
+                        .success(true)
+                        .status(200)
+                        .data(json)
+                        .build();
+            }
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            Map<String, Object> json = new HashMap<>();
+            json.put("message", e.getMessage());
+            json.put("body", null);
+            return Response.builder()
+                    .status(204)
+                    .success(true)
+                    .data(json)
+                    .build();
         }
 
+        return null;
     }
-
-
-
 }
